updated DSL so that it works with the root `Engine` class
