+++ ashley/src/test/kotlin/ktx/ashley/EnginesSpec.kt
@@ -6,8 +6,8 @@ import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
 
-object PooledEnginesSpec : Spek({
-  describe("utilities for pooled engines") {
+object EnginesSpec : Spek({
+  describe("utilities for engines") {
     val engine by memoized {
       PooledEngine()
     }
